fix int overflow

+++ src/snips.js
@@ -4,10 +4,10 @@ snips.cpp = `#include <iostream>
 #include <stdio.h>
 using namespace std;
 
-int N = 1000;   //change value of N to see different status
+long long N = 1000;   //change value of N to see different status
 
 int main() {
-  for(int i = 0; i < N; i++) {
+  for(long long i = 0; i < N; i++) {
     //delay
   }
